// src/app/Models/avaliacao.ts
export class Avaliacao {
  nomepaciente : String;
  atvfisdiaria: String;
  duracaoatvfisdiaria: String;
  maistreshrsdiasendetario: String;
  cholateeindustrialzados: String;
  refri: String;
  frituras: String;
  frutascolridas: String;
  cerealintegral: String;
  qtsporcoescarne: String;
  carneprocessadafreq: String;
  bebida: String;
  temperosindustrializados: String;
  produtosindustrializados: String;
  sal: String;
  suplementos: String;
}